+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/groups/groupDetail/GroupDetailScreen.kt
@@ -148,6 +148,8 @@ fun GroupDetailScreen(
                 initialPage = taggedTabs.indexOfFirst { it.id == initialTabId } + 1,
                 pageCount = { taggedTabs.size + 1 }
             )
+            val isPullToRefreshEnabled = scrollBehavior.state.heightOffset == 0f
+
             Column(
                 modifier = Modifier.padding(top = innerPadding.calculateTopPadding())
             ) {
@@ -161,6 +163,7 @@ fun GroupDetailScreen(
                     taggedTabs = taggedTabs,
                     groupId = groupId,
                     group = group,
+                    isPullToRefreshEnabled = isPullToRefreshEnabled,
                     onTopicClick = onTopicClick,
                     onUserClick = onUserClick,
                     tabContentPadding = PaddingValues(
@@ -454,6 +457,7 @@ fun GroupPager(
     taggedTabs: List<GroupTab>,
     groupId: String,
     group: GroupDetail?,
+    isPullToRefreshEnabled: Boolean,
     onTopicClick: (topicId: String) -> Unit,
     onUserClick: (userId: String) -> Unit,
     modifier: Modifier = Modifier,
@@ -476,6 +480,7 @@ fun GroupPager(
             groupId = groupId,
             tabId = tabId,
             group = group,
+            isPullToRefreshEnabled = isPullToRefreshEnabled,
             onTopicClick = onTopicClick,
             onUserClick = onUserClick,
             contentPadding = tabContentPadding