+++ app/src/main/java/com/github/bumblebee202111/doubean/model/groups/PinnedTabItem.kt
@@ -2,12 +2,11 @@ package com.github.bumblebee202111.doubean.model.groups
 
 import java.util.Calendar
 
-
-data class GroupFavoriteItem(
-    val favoriteDate: Calendar,
+data class PinnedTabItem(
+    val pinnedDate: Calendar,
     val groupId: String,
     val groupName: String?,
     val groupAvatar: String?,
-    val tabId: String? = null,
-    val groupTabName: String? = null,
+    val tabId: String,
+    val tabName: String,
 )